Move _check_inputs_dtype from recurrent to recurrent_internals
